anserv/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
docopt==0.6.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mutagen==1.47.0
openai==0.28.1
pkg_resources==0.0.0
pycryptodomex==3.19.0
requests==2.31.0
SudachiDict-core==20230927
SudachiPy==0.6.7
tqdm==4.66.1
urllib3==2.0.7
websockets==12.0
webvtt-py==0.4.6
yarl==1.9.2
yt-dlp==2023.10.13
